splittet ut logikk for send til beslutter steg og beslutte vedtak steg fra stegservice
